Rick's remarks
