+++ src/styles/app.css
@@ -310,7 +310,7 @@ body {
 }
 
 .headerHome .HomeHero__Nav nav > div > div::before, .headerHome .HomeHero__Nav nav > div > div::after, .headerHome  .HomeHero__Nav nav > div > div {
-  background-color: var(--color-light-green);
+  background-color: var(--color-white);
 }
 
 @media (min-width: 25em) {
@@ -855,9 +855,8 @@ h6 {
 .PlaceOrder:link, .PlaceOrder:visited {
   display: block;
   text-align: center;
-  padding: 4rem 2rem;
-  margin-top: 3rem;
-  margin-bottom: 5rem;
+  padding: 2rem 1rem;
+  margin: 3rem auto 5rem auto;
   color: var(--color-white);
   background-color: var(--color-light-green);
   cursor: pointer;
@@ -865,6 +864,8 @@ h6 {
   position: relative;
   overflow: hidden;
   border-radius: 2rem;
+  width: 50%;
+
 }
 
 .PlaceOrder:link::after {
@@ -876,9 +877,8 @@ h6 {
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url("../images/banner.jpg");
+  background-color: var(--color-dark-green);
   background-size: cover;
-  background-position-y: -35rem;
   opacity: 0.4;
   transition: opacity 0.15s ease-in;
 }
@@ -894,12 +894,15 @@ h6 {
   margin: 0;
   opacity: 1;
   text-transform: uppercase;
-  font-size: 4rem;
+  font-size: 2.5rem;
   text-shadow: 0 1px 3px rgba(0,0,0,0.4);
   z-index: 10;
   position: relative;
 }
 
+.PlaceOrderWidget {
+  text-align: center;
+}
 .ProoForm {
   display: block;
   box-shadow: 0 1px 3px 0 var(--color-box-shadow);
@@ -935,6 +938,17 @@ h6 {
   width: 100%;
 }
 
+.ProoForm--Popup {
+  position: fixed;
+  max-width: 50%;
+  height: 80%;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  overflow: scroll;
+}
+
+
 .mt-0 {
   margin-top: 0;
 }
@@ -946,6 +960,10 @@ h6 {
 }
 
 @media (max-width: 48em) {
+
+.ProoForm__Group {
+  flex-direction: column;
+}
 
   .OrderBox {
     width: 90%;
@@ -981,6 +999,18 @@ h6 {
   grid-template-columns: 27% repeat(2, 1fr) 17%;
   gap: 1.5em;
   padding-top:2rem;
+
+}
+
+.grid-contacts {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 1.5em;
+  padding-top:2rem;
+  width: 100%;
+  font-size: 0.8em;
+  line-height: 1.6em;
+  padding-bottom: 1em;
 }
 
 .footer-address {
@@ -998,8 +1028,12 @@ h6 {
 }
 
 @media (max-width: 48em) {
-  .grid-secondary, .grid-footer {
+  .grid-secondary, .grid-footer, .grid-contacts {
     grid-template-columns: 1fr;
+  }
+
+  .grid-contacts {
+    margin-bottom: 4em;
   }
 
   .footer-address {
